db.jsのハンドリングしている処理をLibrarianサービスにまとめる
